Added feature to return back from sunburst
-Sunburst has a top level for all items.
Users can now return to previous levels in the Sunburst

- improved look of infobox
