+++ assets/colours/graphColours.js
@@ -1,4 +1,4 @@
-export const graphColours = {
+export const GraphColours = {
     "Baby Foods": "#BAABDA",
     "Beverages (Excluding Milks)": "#F8CB2E",
     "Dairy & Plant-Based Beverages": "#39B7E3",
@@ -10,5 +10,6 @@ export const graphColours = {
     "Meat Alternatives": "#482121",
     "Nutritional Beverages & Bars": "#84F1CD",
     "Soups - Sauces - Spices & Other Ingredients": "#005DD0",
-    "Sweets - Sugars & Savoury Snacks": "#824BD0"
+    "Sweets - Sugars & Savoury Snacks": "#824BD0",
+    "All Items": "#808080"
 }
