+++ assets/dimensions/defaultDimensions.js
@@ -1,0 +1,8 @@
+export const DefaultDims = {
+    fontSize: 12,
+    paddingSize: 5,
+    pos: 0,
+    length: 0,
+    borderWidth: 3,
+    lineSpacing: 1
+};
