+++ assets/dimensions/graphDimensions.js
@@ -1,4 +1,4 @@
-export const dims = Object.freeze({
+export const GraphDims = Object.freeze({
     upperGraphWidth: 800,
     upperGraphHeight: 600,
     upperGraphLeft: 60,
@@ -8,7 +8,10 @@ export const dims = Object.freeze({
     upperGraphTooltipMinWidth: 140,
     upperGraphInfoBoxWidth: 240,
     upperGraphInfoBoxHeight: 200,
-    upperGraphInfoBoxFontSize: 12,
+    upperGraphInfoBoxFontSize: 14,
+    upperGraphInfoBoxBorderWidth: 10,
+    upperGraphInfoBoxPaddingLeft: 10,
+    upperGraphInfoBoxLineSpacing: 5,
     upperGraphChartHeadingFontSize: 20,
     lowerGraphWidth: 700,
     lowerGraphHeight: 700,
@@ -22,7 +25,10 @@ export const dims = Object.freeze({
     lowerGraphInfoBoxHeight: 200,
     lowerGraphArcLabelFontSize: 12,
     lowerGraphChartHeadingFontSize: 20,    
-    lowerGraphInfoBoxFontSize: 12,
+    lowerGraphInfoBoxFontSize: 14, 
+    lowerGraphInfoBoxBorderWidth: 10,
+    lowerGraphInfoBoxPaddingLeft: 10,
+    lowerGraphInfoBoxLineSpacing: 10,
     legendFontSize: 12,
     legendSquareSize: 12,
     legendRowHeight: 20,
