+++ data/Food Group descriptions.csv
@@ -88,7 +88,7 @@ Grain Products,,,"Breads and rolls, crackers, breakfast cereals, cakes, cookies 
 Nutritional Beverages & Bars,,,"Energy bars, protein bars, nutrition beverages, meal replacements and protein powders and shakes"
 ,Nutritional Bars,,Energy and protein bars
 ,Meal Replacements & Nutritional Beverages,,"Nutritional beverages, meal replacements (powder and ready to drink), protein powder and shakes"
-Meat &  Poultry,,,"Meat, processed meats and poultry"
+Meat & Poultry,,,"Meat, processed meats and poultry"
 ,Red Meat,,"Beef, lamb, veal, games meats, fresh pork, liver and offal's"
 ,,Beef,Includes ground beef and most cuts of beef. 
 ,,Lamb - Veal & Game Meats,"Includes most cuts of veal, lamb and game meats, includes ground meats. "
