+++ index.html
@@ -17,7 +17,6 @@
               integrity="sha384-50oBUHEmvpQ+1lW4y57PTFmhCaXp0ML5d60M1M7uH2+nqUivzIebhndOJK28anvf" crossorigin="anonymous">
         <link rel="stylesheet" href="./static/GCWeb/css/theme.min.css">
     
-        <link rel="stylesheet" href="./styles/table.css">
         <link rel="stylesheet" href="./styles/styles.css">
         
         <!--<![endif]-->
@@ -281,7 +280,7 @@
                         </label>
                         <select class="form-control" name="ageSexSelect" title="Select age-sex group" id="lowerGraphAgeSexSelect"></select>
                     </div>
-                    <div class="lowerGraphExtra graphExtra">
+                    <div class="graphExtra">
                         <button class="btn btn-info" id="lowerGraphFilterGroupsButton">Filter on Level 2 groups</button>
                     </div>
                     <div class="row">
@@ -655,7 +654,6 @@
 
     <script src="i18next.js"></script>
     <script src="i18nextBrowserLanguageDetector.min.js"></script>
-    <script src="translation.js"></script>
     <script type="module" src="./viewController/foodSourceContributionTool.js"></script>
     <script>
         jQuery.fn.tableToCSV = function () {
