+++ model/model.js
@@ -1,5 +1,6 @@
 import { CSVDataModel } from "./dataModels/csvDataModel.js";
 import { NutrientDataModel } from "./dataModels/nutrientDataModel.js";
+import { FoodGroupDescDataColNames, NutrientDataColNames } from "../assets/strings/columnNames.js";
 
 export class Model {
     #foodGroupDescriptionFileSrc;
@@ -27,11 +28,11 @@ export class Model {
     
         const fullyNestedDataByFoodLevel = Object.freeze(d3.nest()
                                             .key(d => 
-                                                Number.isNaN(d["Food Group Name_Level 3"]) ? 
-                                                    Number.isNaN(d["Food Group Name_Level 2"]) ? 
-                                                        d["Food Group Name_Level 1"]
-                                                        : d["Food Group Name_Level 2"]
-                                                    : d["Food Group Name_Level 3"]
+                                                Number.isNaN(d[FoodGroupDescDataColNames.foodGroupLv3]) ? 
+                                                    Number.isNaN(d[FoodGroupDescDataColNames.foodGroupLv2]) ? 
+                                                        d[FoodGroupDescDataColNames.foodGroupLv1]
+                                                        : d[FoodGroupDescDataColNames.foodGroupLv2]
+                                                    : d[FoodGroupDescDataColNames.foodGroupLv3]
                                             )
                                             .rollup(d => d[0])
                                             .object(data));
@@ -44,27 +45,23 @@ export class Model {
 
         const dataGroupedByNutrientAndDemoList = Object.freeze(d3.nest()
                                         .key(d => d.Nutrient)
-                                        .key(d => d["Age-sex group (*: excludes pregnant or breastfeeding)"])
+                                        .key(d => d[NutrientDataColNames.ageSexGroup])
                                         .object(data));
 
         const dataGroupedByNutrientAndDemo = Object.freeze(d3.nest()
                                         .key(d => d.Nutrient)
-                                        .key(d => d["Age-sex group (*: excludes pregnant or breastfeeding)"])
-                                        .key(d => d["Food group_level1"])
+                                        .key(d => d[NutrientDataColNames.ageSexGroup])
+                                        .key(d => d[NutrientDataColNames.foodGroupLv1])
                                         .object(data));
 
         const fullyNestedDataByFoodGroup = Object.freeze(d3.nest()
                                             .key(d => d.Nutrient)
-                                            .key(d => d["Age-sex group (*: excludes pregnant or breastfeeding)"])
-                                            .key(d => d["Food group_level1"])
-                                            .key(d => Number.isNaN(d["Food group_level2"]) ? "" : d["Food group_level2"] )
-                                            .key(d => Number.isNaN(d["Food group_level3"]) ? "" : d["Food group_level3"] )
+                                            .key(d => d[NutrientDataColNames.ageSexGroup])
+                                            .key(d => d[NutrientDataColNames.foodGroupLv1])
+                                            .key(d => Number.isNaN(d[NutrientDataColNames.foodGroupLv2]) ? "" : d[NutrientDataColNames.foodGroupLv2] )
+                                            .key(d => Number.isNaN(d[NutrientDataColNames.foodGroupLv3]) ? "" : d[NutrientDataColNames.foodGroupLv3] )
                                             .rollup(d => d[0])
                                             .object(data));
-
-        console.log("D1: ", dataGroupedByNutrientAndDemoList);
-        console.log("D2: ", dataGroupedByNutrientAndDemo);
-        console.log("D3: ", fullyNestedDataByFoodGroup);
 
         return new NutrientDataModel(data, dataGroupedByNutrientAndDemoList, dataGroupedByNutrientAndDemo, fullyNestedDataByFoodGroup);
     }
