+++ styles/styles.css
@@ -1,8 +1,5 @@
 .graphExtra {
     margin: 10px 0 0 0;
-}
-
-.lowerGraphExtra {
     float: right;
 }
