+++ viewController/components/component.js
@@ -1,0 +1,8 @@
+
+// Component: Abstract class for building a component used in the UI
+export class Component {
+    // draw(): Draws the overall component
+    draw() {
+
+    }
+}
