+++ viewController/tools/viewTools.js
@@ -1,0 +1,11 @@
+
+
+
+// viewTools: Class of helper functions for view
+export class viewTools {
+
+    /* Returns selected option given a select html tag selector */
+    static getSelector(id) {
+        return d3.select(id).property("value");
+    }
+}
